Improved: The light contribution of indirect (scattered, reflected) light has been improved Fixed: When fog was disabled, rendering transparent textures at the surface of water and lava segments didn't work Fixed: The program could apparently hang when computing lightmaps (it didn't though, it just got unresponsive, but kept working) Fixed: Lightning effects could crash the program Fixed: The update function produced a broken error message when no version info was found on the download servers Changed: Data compression (for stored lightmaps etc.) is active by default now Improved: When downloading an update, D2X-XL will make the program root folder the current directory to make applying the update a bit simpler (no need to direct the sfx archive to the proper folder anymore)
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@14710 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/findfile.h
@@ -7,16 +7,16 @@
 
 // Empty file
 typedef struct FILEFINDSTRUCT {
-	uint32_t size;
+	unsigned long size;
 	char name [256];
 } FILEFINDSTRUCT;
 
-int32_t FileFindFirst (const char *search_str, FILEFINDSTRUCT *ffstruct, int32_t bFindDirs);
-int32_t FileFindNext (FILEFINDSTRUCT *ffstruct, int32_t bFindDirs);
-int32_t FileFindClose (void);
+int FileFindFirst(const char *search_str, FILEFINDSTRUCT *ffstruct, int bFindDirs);
+int FileFindNext(FILEFINDSTRUCT *ffstruct, int bFindDirs);
+int FileFindClose(void);
 
 typedef struct FILETIMESTRUCT {
-	uint16_t date, time;
+	ushort date,time;
 } FILETIMESTRUCT;
 
 #else //!defined (_WIN32)
@@ -37,9 +37,9 @@ typedef struct FILEFINDSTRUCT {
 #define	FF_NORMAL		0
 #define	FF_DIRECTORY	1
 
-int32_t FileFindNext (FILEFINDSTRUCT *ffsP, int32_t nFlags);
-int32_t FileFindFirst (const char *pszFilter, FILEFINDSTRUCT *ffsP, int32_t nFlags);
-void FileFindClose (FILEFINDSTRUCT *ffsP);
+int FileFindNext (FILEFINDSTRUCT *ffsP, int nFlags);
+int FileFindFirst (const char *pszFilter, FILEFINDSTRUCT *ffsP, int nFlags);
+void FileFindClose(FILEFINDSTRUCT *ffsP);
 
 #endif //_WIN32
